Rolling back changes, fixes tests
