+++ localstack/utils/common.py
@@ -27,7 +27,7 @@ import dns.resolver
 from io import BytesIO
 from datetime import datetime, date
 from contextlib import closing
-from six import with_metaclass, iteritems
+from six import with_metaclass
 from six.moves import cStringIO as StringIO
 from six.moves.queue import Queue
 from six.moves.urllib.parse import urlparse, parse_qs
@@ -37,8 +37,6 @@ from localstack.utils import bootstrap
 from localstack.config import DEFAULT_ENCODING
 from localstack.constants import ENV_DEV
 from localstack.utils.bootstrap import FuncThread
-from localstack.utils.aws.aws_models import (ElasticSearch, EventSource,
-    DynamoDB, FirehoseStream, S3Bucket)
 
 # arrays for temporary files and resources
 TMP_FILES = []
@@ -1499,55 +1497,6 @@ def isoformat_milliseconds(t):
         return t.isoformat(timespec='milliseconds')
     except TypeError:
         return t.isoformat()[:-3]
-
-
-def extract_endpoints(code_map, pool={}):
-    result = []
-    identifiers = []
-    for key, code in iteritems(code_map):
-        # Elasticsearch references
-        pattern = r'[\'"](.*\.es\.amazonaws\.com)[\'"]'
-        for es in re.findall(pattern, code):
-            if es not in identifiers:
-                identifiers.append(es)
-                es = EventSource.get(es, pool=pool, type=ElasticSearch)
-                if es:
-                    result.append(es)
-        # Elasticsearch references
-        pattern = r'\.put_record_batch\([^,]+,\s*([^,\s]+)\s*,'
-        for firehose in re.findall(pattern, code):
-            if firehose not in identifiers:
-                identifiers.append(firehose)
-                firehose = EventSource.get(firehose, pool=pool, type=FirehoseStream)
-                if firehose:
-                    result.append(firehose)
-        # DynamoDB references
-        # TODO fix pattern to be generic
-        pattern = r'\.(insert|get)_document\s*\([^,]+,\s*([^,\s]+)\s*,'
-        for (op, dynamo) in re.findall(pattern, code):
-            dynamo = resolve_string_or_variable(dynamo, code_map)
-            if dynamo not in identifiers:
-                identifiers.append(dynamo)
-                dynamo = EventSource.get(dynamo, pool=pool, type=DynamoDB)
-                if dynamo:
-                    result.append(dynamo)
-        # S3 references
-        pattern = r'\.upload_file\([^,]+,\s*([^,\s]+)\s*,'
-        for s3 in re.findall(pattern, code):
-            s3 = resolve_string_or_variable(s3, code_map)
-            if s3 not in identifiers:
-                identifiers.append(s3)
-                s3 = EventSource.get(s3, pool=pool, type=S3Bucket)
-                if s3:
-                    result.append(s3)
-    return result
-
-
-def resolve_string_or_variable(string, code_map):
-    if re.match(r'^["\'].*["\']$', string):
-        return string.replace('"', '').replace("'", '')
-    LOG.warning('Variable resolution not implemented')
-    return None
 
 
 # TODO move to aws_responses.py?
